requirements.txt
docx2txt
extract_msg
fastapi
faiss-cpu
PyPDF2
python-docx
python-dotenv
extract-msg
pytesseract
numpy
scikit-learn
Pillow
langchain
langchain_core
langchain_groq
langchain_huggingface
langchain_openai
langchain_community
sentence_transformers
transformers
pywin32